requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycryptodome==3.23.0
python-dotenv==1.1.0
python-docx
rave-python==1.4.1
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
stripe
gunicorn
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
courlan==1.3.2
dateparser==1.2.1
defusedxml==0.7.1
EbookLib==0.19
fastapi==0.115.12
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
joblib==1.5.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
nltk==3.9.1
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytube==15.0.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
youtube-transcript-api==1.0.3
